+++ src/rct1.c
@@ -25,7 +25,7 @@
 bool rct1_read_sc4(const char *path, rct1_s4 *s4)
 {
 	uint8 *buffer, *decodedBuffer;
-	long length, decodedLength;
+	size_t length, decodedLength;
 	bool success;
 
 	if (!readentirefile(path, (void**)&buffer, (int*)&length)) {
@@ -55,7 +55,7 @@ bool rct1_read_sc4(const char *path, rct1_s4 *s4)
 bool rct1_read_sv4(const char *path, rct1_s4 *s4)
 {
 	uint8 *buffer, *decodedBuffer;
-	long length, decodedLength;
+	size_t length, decodedLength;
 	bool success;
 
 	if (!readentirefile(path, (void**)&buffer, (int*)&length)) {
